Feature/area data type

+++ estate-services/src/main/java/org/egov/ps/model/PropertyDetails.java
@@ -95,8 +95,9 @@ public class PropertyDetails {
 	@JsonProperty("dateOfAuction")
 	private Long dateOfAuction;
 
+	
 	@JsonProperty("areaSqft")
-	private int areaSqft;
+	private BigDecimal areaSqft;
 
 	@JsonProperty("ratePerSqft")
 	private BigDecimal ratePerSqft;
